requirements.txt
streamlit==1.30.0
numpy==1.24.3
pillow
fpdf
pandas
tflite-runtime==2.13.0